+++ src/Traits/InteractsWithRest.php
@@ -3,13 +3,20 @@
 
 namespace TomHart\Restful\Traits;
 
+use Illuminate\Routing\Router;
+use TomHart\Restful\Concerns\HasLinks;
+use TomHart\Restful\LinkBuilder;
+use TomHart\Restful\Routing\Route;
+
 trait InteractsWithRest
 {
     /**
      * @inheritDoc
      */
-    public function getOptionsUrl(): string
+    public function getOptionsRoute(): Route
     {
-        return route($this->getRouteName() . '.options');
+        /** @var HasLinks $this */
+        $links = LinkBuilder::buildLink($this, 'options', app(Router::class));
+        return Route::fromArray($links);
     }
 }